requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
yfinance>=0.2.0
hmmlearn>=0.3.0
ruptures>=1.1.0
pyyaml>=6.0
joblib>=1.2.0
scipy>=1.10.0